repo: delete as-types repo
