--- attestation-service/as-types/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "as-types"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-# TODO: change it to "0.5", once released.
-kbs-types = { git = "https://github.com/virtee/kbs-types", rev = "c90df0e" }
-serde.workspace = true
-serde_json.workspace = true
