+++ attestation-service/attestation-service/Cargo.toml
@@ -20,7 +20,6 @@ rvps-grpc = [ "tonic" ]
 anyhow.workspace = true
 asn1-rs = { version = "0.5.1", optional = true }
 async-trait.workspace = true
-as-types.workspace = true
 az-snp-vtpm = { version = "0.3.0", default-features = false, features = ["verifier"], optional = true }
 base64 = "0.21"
 bincode = "1.3.3"
