+++ attestation-service/attestation-service/src/utils.rs
@@ -4,9 +4,10 @@
 //
 
 use anyhow::*;
-use as_types::TeeEvidenceParsedClaim;
 use serde_json::{Map, Value};
 use serde_variant::to_variant_name;
+
+use crate::verifier::TeeEvidenceParsedClaim;
 
 /// This funciton will transpose the following structured json
 /// ```json
