+++ kbs/src/api/Cargo.toml
@@ -26,7 +26,6 @@ actix-web-httpauth = "0.8.0"
 aes-gcm = { version = "0.10.1", optional = true }
 anyhow.workspace = true
 async-trait.workspace = true
-as-types.workspace = true
 attestation-service = { workspace = true, default-features = false, optional = true }
 base64.workspace = true
 cfg-if.workspace = true
