+++ kbs/src/api/src/attestation/coco/builtin.rs
@@ -5,7 +5,9 @@
 use crate::attestation::Attest;
 use anyhow::*;
 use async_trait::async_trait;
-use attestation_service::{config::Config as AsConfig, AttestationService};
+use attestation_service::{
+    config::Config as AsConfig, policy_engine::SetPolicyInput, AttestationService,
+};
 use kbs_types::Tee;
 
 pub struct Native {
@@ -14,9 +16,12 @@ pub struct Native {
 
 #[async_trait]
 impl Attest for Native {
-    async fn set_policy(&mut self, input: as_types::SetPolicyInput) -> Result<()> {
-        self.inner.set_policy(input).await
+    async fn set_policy(&mut self, input: &[u8]) -> Result<()> {
+        let request: SetPolicyInput =
+            serde_json::from_slice(input).context("parse SetPolicyInput")?;
+        self.inner.set_policy(request).await
     }
+
     async fn verify(&mut self, tee: Tee, nonce: &str, attestation: &str) -> Result<String> {
         self.inner.evaluate(tee, nonce, attestation).await
     }
